requirements.txt
openai-whisper
torch
ffmpeg-python
tqdm
scipy
numpy 